Android fixes

+++ core/formats/mri.cpp
@@ -344,7 +344,7 @@ namespace MR
       write_tag (out, MRI_DATA, 1, is_BE);
       out.put (store_datatype (H.datatype()));
 
-      size_t data_offset = out.tellp();
+      size_t data_offset = int64_t(out.tellp());
       out.close();
 
       std::unique_ptr<ImageIO::Base> io_handler (new ImageIO::Default (H));
